docs/requirements.txt
cigvis
gitpython==3.1.37
vispy
numpy
scipy
matplotlib
plotly
viser
pydata-sphinx-theme==0.13.3
sphinx==6.2.1
sphinxcontrib-apidoc==0.4.0
sphinx-gallery==0.14.0
myst-parser